backend/requirements.txt
flask
flask-cors
nltk
spacy
python-Levenshtein
pdfminer.six
scikit-learn
gunicorn
